maxar renamed to tz3
