+++ modules/3d-tiles/src/index.ts
@@ -2,7 +2,7 @@
 export {Tiles3DLoader} from './tiles-3d-loader';
 export {CesiumIonLoader} from './cesium-ion-loader';
 export {Tile3DSubtreeLoader} from './tile-3d-subtree-loader';
-export {MaxarLoader, MaxarLoaderOptions} from './maxar-loader';
+export {TZ3Loader, TZ3LoaderOptions} from './tz3-loader';
 
 // WRITERS
 export {Tile3DWriter} from './tile-3d-writer';
